Attempt to fix deadlock
Cherry-pick of:
 https://github.com/buildbot/buildbot/pull/3650/commits

Only locks.py related changes are applied. Other files
are for integration testing and were ignored.
There is no guarantee that this fix the Deadlock
described in INF-294.

This is part of a bigger patch applied at mainstream
branch for 1.6.0

Explanation taken from the original patch:

If a build factory has a buildstep with a lock, this will
trigger a deadlock whenever 3 or more builds of this factory
 are started at the same time.
The main reason is that BaseLock uses the == operator to
compare the owners of a given lock.
But if owners are BuildSteps, it doesn't work because
they inherit from ComparableMixin (#d6c6b550),
so two distinct BuildStep objects sharing the same properties
will be considered as equal, although they are
distinct Python objects:
`b0 == b1` but `b0 is not b1`.
We use the `is` operator to fix this.